moved FunctionalityLink to nested class as it is only used from EnableComponentsByFunctionality
