--- Packages/eu.netherlands3d.twin-functionalities/Scripts/FunctionalityLink.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using UnityEngine;
-using UnityEngine.Events;
-using UnityEngine.Serialization;
-
-namespace Netherlands3D.Twin.Functionalities
-{
-    [Serializable]
-    public class FunctionalityLink
-    {
-        [HideInInspector]
-        public string name;
-        public Functionality feature;
-
-        [FormerlySerializedAs("onFeatureToggle")]
-        public UnityEvent<bool> onFeatureToggle = new();
-    }
-}
