--- Packages/eu.netherlands3d.twin-functionalities/Scripts/FunctionalityLink.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6af354fed37449cdb7a0bc697022db6e
-timeCreated: 1685426318
